+++ server/infra/database/CaptureRepository.ts
@@ -91,6 +91,15 @@ export default class CaptureRepository extends BaseRepository<Capture> {
       delete filterObject.reference_id;
     }
 
+    if (filterObject.grower_reference_id) {
+      result.where(
+        `treetracker.grower_account.reference_id`,
+        '=',
+        filterObject.grower_reference_id,
+      );
+      delete filterObject.grower_reference_id;
+    }
+
     if (filterObject.organization_id) {
       result.where(`${this.tableName}.planting_organization_id`, 'in', [
         ...filterObject.organization_id,
@@ -154,7 +163,8 @@ export default class CaptureRepository extends BaseRepository<Capture> {
         `,
         ),
       )
-      .where((builder) => this.filterWhereBuilder(filter, builder));
+      .where((builder) => this.filterWhereBuilder(filter, builder))
+      .distinct();
 
     promise = promise.orderBy(
       sort?.order_by || 'treetracker.capture.created_at',