+++ server/infra/database/SpeciesRepository.ts
@@ -40,6 +40,36 @@ export default class SpeciesRepository extends BaseRepository<Species> {
     return object.rows;
   }
 
+  async countByOrganization(organization_id: number) {
+    const totalSql = `
+    SELECT 
+      species_id as id, total, ts.name, ts.desc
+      FROM 
+      (
+      SELECT 
+      ss.species_id, count(ss.species_id) as total
+      from webmap.species_stat ss
+      WHERE
+      ss.planter_id IN (
+        SELECT
+          id
+        FROM planter p
+        WHERE
+            p.organization_id in ( SELECT entity_id from getEntityRelationshipChildren(${organization_id}))
+      )
+      OR
+        ss.planting_organization_id = ${organization_id}
+      GROUP BY ss.species_id
+      ) s_count
+      JOIN tree_species ts
+      ON ts.id = s_count.species_id
+      ORDER BY total DESC
+
+    `;
+    const total = await this.session.getDB().raw(totalSql);
+    return parseInt(total.rows.length);
+  }
+
   async getByPlanter(planter_id: number, options: FilterOptions) {
     const { limit, offset } = options;
     const sql = `